Require ipcRenderer instead of ipcMain in the rendered context
The following pull request:

https://github.com/jprichardson/electron-mocha/pull/26

introduces a change to make use of electron's new `electron.ipcMain`
property.

However it throws the following error when being used in the renderer
context:

```js
Cannot read property `send` of undefined.
```

The issue is that we should be referring `electron.ipcRenderer` instead
of `electron.ipcMain`, since this code runs in the renderer context.