Add `--interactive` mode
This adds a new `--interactive` option which is similar to `--debug
--renderer` but does not close the Electron window when tests finish
running. Tests can be re-run by reloading the window and all output goes
to the dev tools console rather than the terminal.

This streamlines the workflow when making heavy use of the dev tools
during TDD sessions. Objects such as DOM nodes can be visualized in the
console more easily than they can in the terminal, and the feedback loop
is shorter when you only need to reload the window to re-run tests.

Signed-off-by: Nathan Sobo <nathan@github.com>